backend/requirements.txt
flask==2.2.5
pyyaml==6.0
requests==2.28.2
apscheduler==3.10.1
flask-cors==3.0.10
pytest==7.4.0 